Reformat to module
