+++ packages/discovery-provider/src/api/v1/tracks.py
@@ -47,7 +47,6 @@ from src.queries.generate_unpopulated_trending_tracks import (
     TRENDING_TRACKS_LIMIT,
     TRENDING_TRACKS_TTL_SEC,
 )
-from src.queries.get_comments import get_track_comments, get_track_notification_setting
 from src.queries.get_extended_purchase_gate import get_extended_purchase_gate
 from src.queries.get_feed import get_feed
 from src.queries.get_latest_entities import get_latest_entities
@@ -504,7 +503,7 @@ class TrackComments(Resource):
         args = track_comments_parser.parse_args()
         decoded_id = decode_with_abort(track_id, ns)
         current_user_id = args.get("user_id")
-        track_comments = get_track_comments(args, decoded_id, current_user_id)
+        track_comments = get_track_comment_count(decoded_id, current_user_id)
         return success_response(track_comments)
 
 
@@ -571,7 +570,7 @@ class TrackCommentNotificationSetting(Resource):
         decoded_id = decode_with_abort(track_id, ns)
 
         current_user_id = get_current_user_id(args)
-        track_comments = get_track_notification_setting(decoded_id, current_user_id)
+        track_comments = get_track_comment_count(decoded_id, current_user_id)
         return success_response(track_comments)
