+++ packages/discovery-provider/src/queries/comments/__init__.py
@@ -1,0 +1,22 @@
+from src.queries.comments.replies import get_paginated_replies, get_replies
+from src.queries.comments.track_comments import get_track_comments
+from src.queries.comments.user_comments import get_user_comments
+from src.queries.comments.utils import (
+    COMMENT_REPLIES_DEFAULT_LIMIT,
+    COMMENT_ROOT_DEFAULT_LIMIT,
+    get_is_reacted,
+    get_muted_users,
+    get_track_notification_setting,
+)
+
+__all__ = [
+    "get_track_comments",
+    "get_user_comments",
+    "get_replies",
+    "get_paginated_replies",
+    "get_is_reacted",
+    "get_muted_users",
+    "get_track_notification_setting",
+    "COMMENT_ROOT_DEFAULT_LIMIT",
+    "COMMENT_REPLIES_DEFAULT_LIMIT",
+]
